Access Denied in VP - activeResolvedRequestObject cached and managed in core
