+++ docs/wallet-core/eu.europa.ec.eudi.wallet.issue.openid4vci/-open-id4-vci-manager/index.md
@@ -31,7 +31,8 @@ OpenId4VciManager is the main entry point to issue documents using the OpenId4Vc
 |---|---|
 | [getIssuerMetadata](get-issuer-metadata.md) | [androidJvm]<br>abstract suspend fun [getIssuerMetadata](get-issuer-metadata.md)(): [Result](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin/-result/index.html)&lt;CredentialIssuerMetadata&gt;<br>Provides the issuer metadata |
 | [issueDeferredDocument](issue-deferred-document.md) | [androidJvm]<br>abstract fun [issueDeferredDocument](issue-deferred-document.md)(deferredDocument: DeferredDocument, executor: [Executor](https://developer.android.com/reference/kotlin/java/util/concurrent/Executor.html)? = null, onIssueResult: [OpenId4VciManager.OnDeferredIssueResult](-on-deferred-issue-result/index.md))<br>Issue a deferred document |
-| [issueDocumentByConfigurationIdentifier](issue-document-by-configuration-identifier.md) | [androidJvm]<br>abstract fun [issueDocumentByConfigurationIdentifier](issue-document-by-configuration-identifier.md)(credentialConfigurationId: [String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin/-string/index.html), txCode: [String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin/-string/index.html)? = null, executor: [Executor](https://developer.android.com/reference/kotlin/java/util/concurrent/Executor.html)? = null, onIssueEvent: [OpenId4VciManager.OnIssueEvent](-on-issue-event/index.md))<br>Issue a document using a configuration identifier. |
+| [issueDocumentByConfigurationIdentifier](issue-document-by-configuration-identifier.md) | [androidJvm]<br>abstract fun [~~issueDocumentByConfigurationIdentifier~~](issue-document-by-configuration-identifier.md)(credentialConfigurationId: [String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin/-string/index.html), txCode: [String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin/-string/index.html)? = null, executor: [Executor](https://developer.android.com/reference/kotlin/java/util/concurrent/Executor.html)? = null, onIssueEvent: [OpenId4VciManager.OnIssueEvent](-on-issue-event/index.md))<br>Issue a document using a configuration identifier. |
+| [issueDocumentByConfigurationIdentifiers](issue-document-by-configuration-identifiers.md) | [androidJvm]<br>abstract fun [issueDocumentByConfigurationIdentifiers](issue-document-by-configuration-identifiers.md)(credentialConfigurationIds: [List](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin.collections/-list/index.html)&lt;[String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin/-string/index.html)&gt;, txCode: [String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin/-string/index.html)? = null, executor: [Executor](https://developer.android.com/reference/kotlin/java/util/concurrent/Executor.html)? = null, onIssueEvent: [OpenId4VciManager.OnIssueEvent](-on-issue-event/index.md))<br>Issue a list of documents using a list of configuration identifiers. |
 | [issueDocumentByDocType](issue-document-by-doc-type.md) | [androidJvm]<br>abstract fun [~~issueDocumentByDocType~~](issue-document-by-doc-type.md)(docType: [String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin/-string/index.html), txCode: [String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin/-string/index.html)? = null, executor: [Executor](https://developer.android.com/reference/kotlin/java/util/concurrent/Executor.html)? = null, onIssueEvent: [OpenId4VciManager.OnIssueEvent](-on-issue-event/index.md))<br>Issue a document using a document type |
 | [issueDocumentByFormat](issue-document-by-format.md) | [androidJvm]<br>abstract fun [issueDocumentByFormat](issue-document-by-format.md)(format: DocumentFormat, txCode: [String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin/-string/index.html)? = null, executor: [Executor](https://developer.android.com/reference/kotlin/java/util/concurrent/Executor.html)? = null, onIssueEvent: [OpenId4VciManager.OnIssueEvent](-on-issue-event/index.md))<br>Issue a document using a document format |
 | [issueDocumentByOffer](issue-document-by-offer.md) | [androidJvm]<br>abstract fun [issueDocumentByOffer](issue-document-by-offer.md)(offer: [Offer](../-offer/index.md), txCode: [String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin/-string/index.html)? = null, executor: [Executor](https://developer.android.com/reference/kotlin/java/util/concurrent/Executor.html)? = null, onIssueEvent: [OpenId4VciManager.OnIssueEvent](-on-issue-event/index.md))<br>Issue a document using an offer |
