+++ docs/wallet-core/eu.europa.ec.eudi.wallet.issue.openid4vci/-open-id4-vci-manager/issue-document-by-configuration-identifier.md
@@ -3,7 +3,15 @@
 # issueDocumentByConfigurationIdentifier
 
 [androidJvm]\
-abstract fun [issueDocumentByConfigurationIdentifier](issue-document-by-configuration-identifier.md)(credentialConfigurationId: [String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin/-string/index.html), txCode: [String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin/-string/index.html)? = null, executor: [Executor](https://developer.android.com/reference/kotlin/java/util/concurrent/Executor.html)? = null, onIssueEvent: [OpenId4VciManager.OnIssueEvent](-on-issue-event/index.md))
+abstract fun [~~issueDocumentByConfigurationIdentifier~~](issue-document-by-configuration-identifier.md)(credentialConfigurationId: [String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin/-string/index.html), txCode: [String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin/-string/index.html)? = null, executor: [Executor](https://developer.android.com/reference/kotlin/java/util/concurrent/Executor.html)? = null, onIssueEvent: [OpenId4VciManager.OnIssueEvent](-on-issue-event/index.md))
+
+---
+
+### Deprecated
+
+Use issueDocumentByConfigurationIdentifiers that accepts a list of identifiers
+
+---
 
 Issue a document using a configuration identifier.
