+++ docs/wallet-core/eu.europa.ec.eudi.wallet.presentation/-presentation-manager-impl/index.md
@@ -33,6 +33,7 @@ Implementation of the [PresentationManager](../-presentation-manager/index.md) i
 | [addTransferEventListener](add-transfer-event-listener.md) | [androidJvm]<br>open override fun [addTransferEventListener](add-transfer-event-listener.md)(listener: TransferEvent.Listener): &lt;Error class: unknown class&gt; |
 | [disableNFCEngagement](disable-n-f-c-engagement.md) | [androidJvm]<br>open override fun [disableNFCEngagement](disable-n-f-c-engagement.md)(activity: [ComponentActivity](https://developer.android.com/reference/kotlin/androidx/activity/ComponentActivity.html)): &lt;Error class: unknown class&gt;<br>Disable the NFC device engagement for the wallet. This method should be called in the activity's [ComponentActivity.onPause](https://developer.android.com/reference/kotlin/androidx/activity/ComponentActivity.html#onpause) method. |
 | [enableNFCEngagement](enable-n-f-c-engagement.md) | [androidJvm]<br>open override fun [enableNFCEngagement](enable-n-f-c-engagement.md)(activity: [ComponentActivity](https://developer.android.com/reference/kotlin/androidx/activity/ComponentActivity.html)): &lt;Error class: unknown class&gt;<br>Enable the NFC device engagement for the wallet. This method should be called in the activity's [ComponentActivity.onResume](https://developer.android.com/reference/kotlin/androidx/activity/ComponentActivity.html#onresume) method. |
+| [rejectRemotePresentation](reject-remote-presentation.md) | [androidJvm]<br>open override fun [rejectRemotePresentation](reject-remote-presentation.md)()<br>Rejects the current remote presentation request. |
 | [removeAllTransferEventListeners](remove-all-transfer-event-listeners.md) | [androidJvm]<br>open override fun [removeAllTransferEventListeners](remove-all-transfer-event-listeners.md)(): &lt;Error class: unknown class&gt; |
 | [removeTransferEventListener](remove-transfer-event-listener.md) | [androidJvm]<br>open override fun [removeTransferEventListener](remove-transfer-event-listener.md)(listener: TransferEvent.Listener): &lt;Error class: unknown class&gt; |
 | [sendResponse](send-response.md) | [androidJvm]<br>open override fun [sendResponse](send-response.md)(response: Response)<br>Send a response to verifier |
