+++ docs/wallet-core/eu.europa.ec.eudi.wallet.presentation/-presentation-manager/index.md
@@ -40,6 +40,7 @@ It provides also functionality to start the NFC engagement by calling [enableNFC
 | [addTransferEventListener](../../eu.europa.ec.eudi.wallet.transactionLogging.presentation/-transactions-decorator/index.md#-1321081126%2FFunctions%2F1615067946) | [androidJvm]<br>abstract fun [addTransferEventListener](../../eu.europa.ec.eudi.wallet.transactionLogging.presentation/-transactions-decorator/index.md#-1321081126%2FFunctions%2F1615067946)(listener: TransferEvent.Listener): TransferEvent.Listenable |
 | [disableNFCEngagement](disable-n-f-c-engagement.md) | [androidJvm]<br>abstract fun [disableNFCEngagement](disable-n-f-c-engagement.md)(activity: [ComponentActivity](https://developer.android.com/reference/kotlin/androidx/activity/ComponentActivity.html)): [PresentationManager](index.md)<br>Disable the NFC device engagement for the wallet. This method should be called in the activity's [ComponentActivity.onPause](https://developer.android.com/reference/kotlin/androidx/activity/ComponentActivity.html#onpause) method. |
 | [enableNFCEngagement](enable-n-f-c-engagement.md) | [androidJvm]<br>abstract fun [enableNFCEngagement](enable-n-f-c-engagement.md)(activity: [ComponentActivity](https://developer.android.com/reference/kotlin/androidx/activity/ComponentActivity.html)): [PresentationManager](index.md)<br>Enable the NFC device engagement for the wallet. This method should be called in the activity's [ComponentActivity.onResume](https://developer.android.com/reference/kotlin/androidx/activity/ComponentActivity.html#onresume) method. |
+| [rejectRemotePresentation](reject-remote-presentation.md) | [androidJvm]<br>abstract fun [rejectRemotePresentation](reject-remote-presentation.md)()<br>Rejects the current remote presentation request. |
 | [removeAllTransferEventListeners](index.md#-812467576%2FFunctions%2F1615067946) | [androidJvm]<br>abstract fun [removeAllTransferEventListeners](index.md#-812467576%2FFunctions%2F1615067946)(): TransferEvent.Listenable |
 | [removeTransferEventListener](../../eu.europa.ec.eudi.wallet.transactionLogging.presentation/-transactions-decorator/index.md#1314902509%2FFunctions%2F1615067946) | [androidJvm]<br>abstract fun [removeTransferEventListener](../../eu.europa.ec.eudi.wallet.transactionLogging.presentation/-transactions-decorator/index.md#1314902509%2FFunctions%2F1615067946)(listener: TransferEvent.Listener): TransferEvent.Listenable |
 | [sendResponse](send-response.md) | [androidJvm]<br>abstract fun [sendResponse](send-response.md)(response: Response)<br>Send a response to verifier |
