+++ docs/wallet-core/eu.europa.ec.eudi.wallet.transactionLogging.presentation.parsing/index.md
@@ -6,7 +6,7 @@
 
 | Name | Summary |
 |---|---|
-| [claims](claims.md) | [androidJvm]<br>val SdJwt&lt;JwtAndClaims&gt;.[claims](claims.md): [Map](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin.collections/-map/index.html)&lt;[List](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin.collections/-list/index.html)&lt;[String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin/-string/index.html)&gt;, JsonElement?&gt;<br>This extension property is used to get the claims from the sd-jwt object and get a map of claims with the path as the key and the value as the value It uses the DefaultSdJwtOps.recreateClaimsAndDisclosuresPerClaim method and then selects the claims from the json object using the path |
+| [claims](claims.md) | [androidJvm]<br>val SdJwt&lt;&lt;Error class: unknown class&gt;&gt;.[claims](claims.md): [Map](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin.collections/-map/index.html)&lt;[List](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin.collections/-list/index.html)&lt;[String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin/-string/index.html)&gt;, JsonElement?&gt;<br>This extension property is used to get the claims from the sd-jwt object and get a map of claims with the path as the key and the value as the value It uses the DefaultSdJwtOps.recreateClaimsAndDisclosuresPerClaim method and then selects the claims from the json object using the path |
 | [valueWithoutKeyBinding](value-without-key-binding.md) | [androidJvm]<br>val VerifiablePresentation.Generic.[valueWithoutKeyBinding](value-without-key-binding.md): [String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin/-string/index.html)<br>This function is used to remove the key binding from the sd-jwt due to limitations of the sd-jwt library and DefaultSdJwtOps.verify method. |
 
 ## Functions
@@ -15,7 +15,7 @@
 |---|---|
 | [findClaimMetadataForSdJwtVc](find-claim-metadata-for-sd-jwt-vc.md) | [androidJvm]<br>fun [findClaimMetadataForSdJwtVc](find-claim-metadata-for-sd-jwt-vc.md)(path: [List](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin.collections/-list/index.html)&lt;[String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin/-string/index.html)&gt;, metadata: IssuerMetadata?): IssuerMetadata.Claim?<br>Function to find the claim metadata from the path and metadata It takes the path as a list of strings and metadata and iterates IssuerMetadata.claims to find the claim metadata |
 | [getPresentedDocumentsFromClaims](get-presented-documents-from-claims.md) | [androidJvm]<br>fun [getPresentedDocumentsFromClaims](get-presented-documents-from-claims.md)(claims: [Map](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin.collections/-map/index.html)&lt;[List](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin.collections/-list/index.html)&lt;[String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin/-string/index.html)&gt;, JsonElement?&gt;, metadata: IssuerMetadata?): [PresentedDocument](../eu.europa.ec.eudi.wallet.transactionLogging.presentation/-presented-document/index.md)<br>Function to get the presented documents from the claims It takes the claims as a map of path to value and metadata and returns a PresentedDocument object |
-| [getSdJwt](get-sd-jwt.md) | [androidJvm]<br>fun [getSdJwt](get-sd-jwt.md)(sdJwt: [String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin/-string/index.html)): SdJwt&lt;JwtAndClaims&gt;?<br>Function to parse the sd-jwt string and return the SdJwt object |
+| [getSdJwt](get-sd-jwt.md) | [androidJvm]<br>fun [getSdJwt](get-sd-jwt.md)(sdJwt: [String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin/-string/index.html)): SdJwt&lt;&lt;Error class: unknown class&gt;&gt;?<br>Function to parse the sd-jwt string and return the SdJwt object |
 | [parseMsoMdoc](parse-mso-mdoc.md) | [androidJvm]<br>fun [parseMsoMdoc](parse-mso-mdoc.md)(rawResponse: [ByteArray](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin/-byte-array/index.html), sessionTranscript: [ByteArray](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin/-byte-array/index.html)?, metadata: [List](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin.collections/-list/index.html)&lt;[String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin/-string/index.html)&gt;): [List](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin.collections/-list/index.html)&lt;[PresentedDocument](../eu.europa.ec.eudi.wallet.transactionLogging.presentation/-presented-document/index.md)&gt;<br>Parses the MSO mdoc response and returns a list of presented documents. |
 | [parseMsoMdocFromVp](parse-mso-mdoc-from-vp.md) | [androidJvm]<br>fun [parseMsoMdocFromVp](parse-mso-mdoc-from-vp.md)(vp: VerifiablePresentation.Generic, metadata: [TransactionLog.Metadata](../eu.europa.ec.eudi.wallet.transactionLogging/-transaction-log/-metadata/index.md)): [PresentedDocument](../eu.europa.ec.eudi.wallet.transactionLogging.presentation/-presented-document/index.md)?<br>Parses an mso_mdoc document from a Verifiable Presentation. This function extracts the mso_mdoc data from the verifiable presentation, decodes it from Base64, and then passes it to the parseMsoMdoc function. |
 | [parsePresentationTransactionLog](parse-presentation-transaction-log.md) | [androidJvm]<br>fun [parsePresentationTransactionLog](parse-presentation-transaction-log.md)(transactionLog: [TransactionLog](../eu.europa.ec.eudi.wallet.transactionLogging/-transaction-log/index.md)): [PresentationTransactionLog](../eu.europa.ec.eudi.wallet.transactionLogging.presentation/-presentation-transaction-log/index.md)<br>Parses a [TransactionLog](../eu.europa.ec.eudi.wallet.transactionLogging/-transaction-log/index.md) and returns a [PresentationTransactionLog](../eu.europa.ec.eudi.wallet.transactionLogging.presentation/-presentation-transaction-log/index.md). |
