+++ docs/wallet-core/eu.europa.ec.eudi.wallet.transactionLogging.presentation/-transactions-decorator/index.md
@@ -38,6 +38,7 @@ androidJvm
 | [addTransferEventListener](index.md#-1321081126%2FFunctions%2F1615067946) | [androidJvm]<br>open override fun [addTransferEventListener](index.md#-1321081126%2FFunctions%2F1615067946)(listener: TransferEvent.Listener): TransferEvent.Listenable |
 | [disableNFCEngagement](../../eu.europa.ec.eudi.wallet.presentation/-presentation-manager/disable-n-f-c-engagement.md) | [androidJvm]<br>open override fun [disableNFCEngagement](../../eu.europa.ec.eudi.wallet.presentation/-presentation-manager/disable-n-f-c-engagement.md)(activity: [ComponentActivity](https://developer.android.com/reference/kotlin/androidx/activity/ComponentActivity.html)): [PresentationManager](../../eu.europa.ec.eudi.wallet.presentation/-presentation-manager/index.md)<br>Disable the NFC device engagement for the wallet. This method should be called in the activity's ComponentActivity.onPause method. |
 | [enableNFCEngagement](../../eu.europa.ec.eudi.wallet.presentation/-presentation-manager/enable-n-f-c-engagement.md) | [androidJvm]<br>open override fun [enableNFCEngagement](../../eu.europa.ec.eudi.wallet.presentation/-presentation-manager/enable-n-f-c-engagement.md)(activity: [ComponentActivity](https://developer.android.com/reference/kotlin/androidx/activity/ComponentActivity.html)): [PresentationManager](../../eu.europa.ec.eudi.wallet.presentation/-presentation-manager/index.md)<br>Enable the NFC device engagement for the wallet. This method should be called in the activity's ComponentActivity.onResume method. |
+| [rejectRemotePresentation](../../eu.europa.ec.eudi.wallet.presentation/-presentation-manager/reject-remote-presentation.md) | [androidJvm]<br>open override fun [rejectRemotePresentation](../../eu.europa.ec.eudi.wallet.presentation/-presentation-manager/reject-remote-presentation.md)()<br>Rejects the current remote presentation request. |
 | [removeAllTransferEventListeners](remove-all-transfer-event-listeners.md) | [androidJvm]<br>open override fun [removeAllTransferEventListeners](remove-all-transfer-event-listeners.md)(): &lt;Error class: unknown class&gt;<br>Removes all transfer event listeners from the delegate [PresentationManager](../../eu.europa.ec.eudi.wallet.presentation/-presentation-manager/index.md) and then re-adds the internal transactionListener. |
 | [removeTransferEventListener](index.md#1314902509%2FFunctions%2F1615067946) | [androidJvm]<br>open override fun [removeTransferEventListener](index.md#1314902509%2FFunctions%2F1615067946)(listener: TransferEvent.Listener): TransferEvent.Listenable |
 | [sendResponse](send-response.md) | [androidJvm]<br>open override fun [sendResponse](send-response.md)(response: Response)<br>Sends a response using the delegate [PresentationManager](../../eu.europa.ec.eudi.wallet.presentation/-presentation-manager/index.md) and logs the outcome. |
