+++ docs/wallet-core/eu.europa.ec.eudi.wallet.transfer.openId4vp/-open-id4-vp-manager/index.md
@@ -37,6 +37,7 @@ This class is responsible for configuring, initializing, and orchestrating the O
 | Name | Summary |
 |---|---|
 | [addTransferEventListener](add-transfer-event-listener.md) | [androidJvm]<br>open override fun [addTransferEventListener](add-transfer-event-listener.md)(listener: TransferEvent.Listener): &lt;Error class: unknown class&gt;<br>Registers a new transfer event listener. |
+| [reject](reject.md) | [androidJvm]<br>fun [reject](reject.md)()<br>Called when the USER cancels-rejects the transaction from the UI. Uses the cached resolved request object to send the rejection. |
 | [removeAllTransferEventListeners](remove-all-transfer-event-listeners.md) | [androidJvm]<br>open override fun [removeAllTransferEventListeners](remove-all-transfer-event-listeners.md)(): &lt;Error class: unknown class&gt;<br>Removes all transfer event listeners. |
 | [removeTransferEventListener](remove-transfer-event-listener.md) | [androidJvm]<br>open override fun [removeTransferEventListener](remove-transfer-event-listener.md)(listener: TransferEvent.Listener): &lt;Error class: unknown class&gt;<br>Removes a transfer event listener. |
 | [resolveRequestUri](resolve-request-uri.md) | [androidJvm]<br>fun [resolveRequestUri](resolve-request-uri.md)(uri: [String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin/-string/index.html))<br>Resolves a request URI. This method is asynchronous and the result is emitted through the TransferEvent.Listener interface. Every time it is called it cancels any previous request that is being resolved. This will lead to the TransferEvent.Disconnected event being emitted. |
