+++ docs/wallet-core/eu.europa.ec.eudi.wallet.transfer.openId4vp/-open-id4-vp-manager/reject.md
@@ -1,0 +1,8 @@
+//[wallet-core](../../../index.md)/[eu.europa.ec.eudi.wallet.transfer.openId4vp](../index.md)/[OpenId4VpManager](index.md)/[reject](reject.md)
+
+# reject
+
+[androidJvm]\
+fun [reject](reject.md)()
+
+Called when the USER cancels-rejects the transaction from the UI. Uses the cached resolved request object to send the rejection.
