+++ docs/wallet-core/eu.europa.ec.eudi.wallet.transfer.openId4vp/-open-id4-vp-request/response-encryption-parameters.md
@@ -1,0 +1,8 @@
+//[wallet-core](../../../index.md)/[eu.europa.ec.eudi.wallet.transfer.openId4vp](../index.md)/[OpenId4VpRequest](index.md)/[responseEncryptionParameters](response-encryption-parameters.md)
+
+# responseEncryptionParameters
+
+[androidJvm]\
+val [responseEncryptionParameters](response-encryption-parameters.md): EncryptionParameters?
+
+Computes the encryption parameters required to send a response (Success or Error) if the Verifier requested JARM encryption.
