+++ docs/wallet-core/eu.europa.ec.eudi.wallet/-eudi-wallet-impl/index.md
@@ -43,6 +43,7 @@ Implementation of [EudiWallet](../-eudi-wallet/index.md)
 | [getDocumentById](index.md#-233386509%2FFunctions%2F1615067946) | [androidJvm]<br>open override fun [getDocumentById](index.md#-233386509%2FFunctions%2F1615067946)(documentId: DocumentId): Document? |
 | [getDocuments](index.md#-1741352450%2FFunctions%2F1615067946) | [androidJvm]<br>open override fun [getDocuments](index.md#-1741352450%2FFunctions%2F1615067946)(predicate: (Document) -&gt; [Boolean](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin/-boolean/index.html)?): [List](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin.collections/-list/index.html)&lt;Document&gt; |
 | [loadMdocSampleDocuments](index.md#408834997%2FFunctions%2F1615067946) | [androidJvm]<br>open override fun [loadMdocSampleDocuments](index.md#408834997%2FFunctions%2F1615067946)(sampleData: [ByteArray](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin/-byte-array/index.html), createSettings: CreateDocumentSettings, documentNamesMap: [Map](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin.collections/-map/index.html)&lt;DocType, [String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin/-string/index.html)&gt;?): Outcome&lt;[List](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin.collections/-list/index.html)&lt;DocumentId&gt;&gt; |
+| [rejectRemotePresentation](../../eu.europa.ec.eudi.wallet.presentation/-presentation-manager/reject-remote-presentation.md) | [androidJvm]<br>open override fun [rejectRemotePresentation](../../eu.europa.ec.eudi.wallet.presentation/-presentation-manager/reject-remote-presentation.md)()<br>Rejects the current remote presentation request. |
 | [removeAllTransferEventListeners](../../eu.europa.ec.eudi.wallet.presentation/-presentation-manager/index.md#-812467576%2FFunctions%2F1615067946) | [androidJvm]<br>open override fun [removeAllTransferEventListeners](../../eu.europa.ec.eudi.wallet.presentation/-presentation-manager/index.md#-812467576%2FFunctions%2F1615067946)(): TransferEvent.Listenable |
 | [removeTransferEventListener](../../eu.europa.ec.eudi.wallet.transactionLogging.presentation/-transactions-decorator/index.md#1314902509%2FFunctions%2F1615067946) | [androidJvm]<br>open override fun [removeTransferEventListener](../../eu.europa.ec.eudi.wallet.transactionLogging.presentation/-transactions-decorator/index.md#1314902509%2FFunctions%2F1615067946)(listener: TransferEvent.Listener): TransferEvent.Listenable |
 | [resolveStatus](../../eu.europa.ec.eudi.wallet.statium/-document-status-resolver/resolve-status.md) | [androidJvm]<br>open suspend override fun [resolveStatus](../../eu.europa.ec.eudi.wallet.statium/-document-status-resolver/resolve-status.md)(document: IssuedDocument): [Result](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin-stdlib/kotlin/-result/index.html)&lt;Status&gt;<br>Resolves the status of the given document |
