+++ wallet-core/src/main/java/eu/europa/ec/eudi/wallet/presentation/PresentationManager.kt
@@ -91,6 +91,20 @@ interface PresentationManager : TransferEvent.Listenable, ReaderTrustStoreAware 
      */
     fun stopRemotePresentation()
 
+    /**
+     * Rejects the current remote presentation request.
+     *
+     * This method should be called when the end-user explicitly declines to share their credentials
+     * (e.g., by clicking a "Cancel" or "Reject" button in the UI).
+     *
+     * It triggers the following actions:
+     * Retrieves the cached request details from the current session.
+     * Sends a "Negative Consensus" (access_denied) response to the Verifier, ensuring the
+     * transaction is closed cleanly on the server side.
+     * Clears the internal state and stops the presentation flow.
+     *
+     * Note: If no active remote presentation is in progress, this method will log a warning and return safely.
+     */
     fun rejectRemotePresentation()
 
     /**
